Refactor HawkSearcher insertion pipeline
Modified `search_to_insert`, `insert_from_search_results`, and `search_init`
function calls to be more clear about handling of entry points, and to remove
redundant input of `insertion_layer` from function signature of
`insert_from_search_results`.  Now the search results contain a list of graph
neighborhoods representing the nearest neighbors of the inserted vector at all
layers in which the vector is to be inserted, along with a flag representing
whether the inserted vector is to be set as the entry point.  This holds
whether or not the entry point has been initialized; in the latter case the
inserted vector is set as the entry point, with empty neighbor sets.

Added doc strings on some function calls to more clearly specify this behavior.