Hawk Searcher Refactor (#33)
* Reorganize module structure

* Remove EntryPoint struct from public interface of GraphStore to simplify abstractions

* Refactor and document HawkSearcher insertion pipeline

Modified `search_to_insert`, `insert_from_search_results`, and `search_init`
function calls to be more clear about handling of entry points, and to remove
redundant input of `insertion_layer` from function signature of
`insert_from_search_results`.  Now the search results contain a list of graph
neighborhoods representing the nearest neighbors of the inserted vector at all
layers in which the vector is to be inserted, along with a flag representing
whether the inserted vector is to be set as the entry point.  This holds
whether or not the entry point has been initialized; in the latter case the
inserted vector is set as the entry point, with empty neighbor sets.

* Add convenience search function to HawkSearcher to execute full insertion pipeline

---------

Co-authored-by: Bryan Gillespie <bryan@inversed.tech>